refactor: split readers into their own files
