+++ src/composed/message/types.rs
@@ -1,9 +1,6 @@
 use std::io::{self, BufRead, Read};
 
-use bytes::{Buf, Bytes, BytesMut};
-#[cfg(feature = "bzip2")]
-use bzip2::write::BzEncoder;
-use digest::DynDigest;
+use bytes::{Buf, Bytes};
 use log::{debug, warn};
 
 use crate::armor;
@@ -11,19 +8,15 @@ use crate::composed::message::decrypt::*;
 use crate::composed::signed_key::SignedSecretKey;
 use crate::errors::{Error, Result};
 use crate::packet::{
-    LiteralDataHeader, OnePassSignature, OpsVersionSpecific, Packet, PacketHeader, PacketTrait,
-    PublicKeyEncryptedSessionKey, Signature, SignatureVersionSpecific, SymKeyEncryptedSessionKey,
+    LiteralDataHeader, OnePassSignature, Packet, PacketHeader, PacketTrait,
+    PublicKeyEncryptedSessionKey, Signature, SymKeyEncryptedSessionKey,
 };
 use crate::ser::Serialize;
-use crate::types::{
-    CompressionAlgorithm, EskType, Fingerprint, KeyDetails, KeyId, KeyVersion, PacketLength,
-    Password, PkeskVersion, PublicKeyTrait, SecretKeyTrait, StringToKey, Tag,
-};
-use crate::util::fill_buffer;
+use crate::types::{EskType, Fingerprint, KeyDetails, Password, PkeskVersion, PublicKeyTrait, Tag};
 
 use super::reader::{
-    CompressedDataReader, LiteralDataReader, PacketBodyReader, SymEncryptedDataReader,
-    SymEncryptedProtectedDataReader,
+    CompressedDataReader, LiteralDataReader, PacketBodyReader, SignatureBodyReader,
+    SignatureOnePassReader, SymEncryptedDataReader, SymEncryptedProtectedDataReader,
 };
 
 /// An [OpenPGP message](https://www.rfc-editor.org/rfc/rfc9580.html#name-openpgp-messages)
@@ -58,451 +51,7 @@ pub enum Message<'a> {
     },
 }
 
-#[derive(derive_more::Debug)]
-pub enum SignatureBodyReader<'a> {
-    Init {
-        /// Running hasher
-        #[debug("hasher")]
-        hasher: Box<dyn DynDigest>,
-        /// Data source
-        source: Box<Message<'a>>,
-    },
-    Body {
-        /// Running hasher
-        #[debug("hasher")]
-        hasher: Box<dyn DynDigest>,
-        /// Data source
-        source: Box<Message<'a>>,
-        buffer: BytesMut,
-    },
-    Done {
-        /// Finalized hash
-        hash: Box<[u8]>,
-        /// Data source
-        source: Box<Message<'a>>,
-    },
-    Error,
-}
-
-impl<'a> SignatureBodyReader<'a> {
-    pub(crate) fn new(sig: &Signature, source: Box<Message<'a>>) -> Result<Self> {
-        let mut hasher = sig.config.hash_alg.new_hasher()?;
-        if let SignatureVersionSpecific::V6 { ref salt, .. } = sig.config.version_specific {
-            // Salt size must match the expected length for the hash algorithm that is used
-            //
-            // See: https://www.rfc-editor.org/rfc/rfc9580.html#section-5.2.3-2.10.2.1.1
-            ensure_eq!(
-                sig.config.hash_alg.salt_len(),
-                Some(salt.len()),
-                "Illegal salt length {} for a V6 Signature using {:?}",
-                salt.len(),
-                sig.config.hash_alg,
-            );
-
-            hasher.update(salt.as_ref());
-        }
-
-        Ok(Self::Init { hasher, source })
-    }
-
-    pub fn hash(&self) -> Option<&[u8]> {
-        match self {
-            Self::Done { hash, .. } => Some(&hash),
-            Self::Error => panic!("error state"),
-            _ => None,
-        }
-    }
-
-    pub fn get_ref(&self) -> &Box<Message<'a>> {
-        match self {
-            Self::Init { source, .. } => source,
-            Self::Body { source, .. } => source,
-            Self::Done { source, .. } => source,
-            Self::Error => panic!("error state"),
-        }
-    }
-
-    pub fn into_inner(self) -> PacketBodyReader<Box<dyn BufRead + 'a>> {
-        match self {
-            Self::Init { source, .. } => source.into_inner(),
-            Self::Body { source, .. } => source.into_inner(),
-            Self::Done { source, .. } => source.into_inner(),
-            Self::Error => panic!("error state"),
-        }
-    }
-
-    fn fill_inner(&mut self) -> io::Result<()> {
-        todo!()
-    }
-
-    pub fn is_done(&self) -> bool {
-        matches!(self, Self::Done { .. })
-    }
-
-    fn decompress(self) -> Result<Self> {
-        match self {
-            Self::Init { hasher, source } => {
-                let source = source.decompress()?;
-                Ok(Self::Init {
-                    hasher,
-                    source: Box::new(source),
-                })
-            }
-            _ => {
-                bail!("cannot decompress message that has already been read from");
-            }
-        }
-    }
-
-    fn decrypt(
-        self,
-        key_pws: &[Password],
-        keys: &[&SignedSecretKey],
-    ) -> Result<(Self, Vec<Fingerprint>)> {
-        match self {
-            Self::Init { hasher, source } => {
-                let (source, fps) = source.decrypt(key_pws, keys)?;
-                Ok((
-                    Self::Init {
-                        hasher,
-                        source: Box::new(source),
-                    },
-                    fps,
-                ))
-            }
-            _ => {
-                bail!("cannot decrypt message that has already been read from");
-            }
-        }
-    }
-
-    fn decrypt_with_password(self, msg_pw: &Password) -> Result<Self> {
-        match self {
-            Self::Init { hasher, source } => {
-                let source = source.decrypt_with_password(msg_pw)?;
-                Ok(Self::Init {
-                    hasher,
-                    source: Box::new(source),
-                })
-            }
-            _ => {
-                bail!("cannot decrypt message that has already been read from");
-            }
-        }
-    }
-}
-
-impl BufRead for SignatureBodyReader<'_> {
-    fn fill_buf(&mut self) -> io::Result<&[u8]> {
-        self.fill_inner()?;
-        match self {
-            Self::Init { .. } => panic!("invalid state"),
-            Self::Body { buffer, .. } => Ok(&buffer[..]),
-            Self::Done { .. } => Ok(&[][..]),
-            Self::Error => panic!("error state"),
-        }
-    }
-
-    fn consume(&mut self, amt: usize) {
-        match self {
-            Self::Init { .. } => panic!("invalid state"),
-            Self::Body { buffer, .. } => {
-                buffer.advance(amt);
-            }
-            Self::Done { .. } => {}
-            Self::Error => panic!("error state"),
-        }
-    }
-}
-
-impl Read for SignatureBodyReader<'_> {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        self.fill_inner()?;
-        match self {
-            Self::Init { .. } => panic!("invalid state"),
-            Self::Body { buffer, .. } => {
-                let to_write = buffer.remaining().min(buf.len());
-                buffer.copy_to_slice(&mut buf[..to_write]);
-                Ok(to_write)
-            }
-            Self::Done { .. } => Ok(0),
-            Self::Error => panic!("error state"),
-        }
-    }
-}
-
-#[derive(derive_more::Debug)]
-pub enum SignatureOnePassReader<'a> {
-    Init {
-        /// Running hasher
-        #[debug("hasher")]
-        hasher: Box<dyn DynDigest>,
-        /// Data source
-        source: Box<Message<'a>>,
-    },
-    Body {
-        /// Running hasher
-        #[debug("hasher")]
-        hasher: Box<dyn DynDigest>,
-        /// Data source
-        source: Box<Message<'a>>,
-        buffer: BytesMut,
-    },
-    Done {
-        /// Finalized hash
-        hash: Box<[u8]>,
-        /// Data source
-        source: Box<Message<'a>>,
-        /// Final signature,
-        signature: Signature,
-    },
-    Error,
-}
-
-impl<'a> SignatureOnePassReader<'a> {
-    pub(crate) fn new(ops: &OnePassSignature, source: Box<Message<'a>>) -> Result<Self> {
-        let mut hasher = ops.hash_algorithm().new_hasher()?;
-        if let OpsVersionSpecific::V6 { salt, .. } = ops.version_specific() {
-            // Salt size must match the expected length for the hash algorithm that is used
-            //
-            // See: https://www.rfc-editor.org/rfc/rfc9580.html#section-5.2.3-2.10.2.1.1
-            ensure_eq!(
-                ops.hash_algorithm().salt_len(),
-                Some(salt.len()),
-                "Illegal salt length {} for a V6 Signature using {:?}",
-                salt.len(),
-                ops.hash_algorithm(),
-            );
-
-            hasher.update(salt.as_ref());
-        }
-
-        Ok(Self::Init { hasher, source })
-    }
-
-    pub fn hash(&self) -> Option<&[u8]> {
-        match self {
-            Self::Done { hash, .. } => Some(&hash),
-            Self::Error => panic!("error state"),
-            _ => None,
-        }
-    }
-
-    pub fn signature(&self) -> Option<&Signature> {
-        match self {
-            Self::Done { signature, .. } => Some(signature),
-            Self::Error => panic!("error state"),
-            _ => None,
-        }
-    }
-
-    pub fn get_ref(&self) -> &Box<Message<'a>> {
-        match self {
-            Self::Init { source, .. } => source,
-            Self::Body { source, .. } => source,
-            Self::Done { source, .. } => source,
-            Self::Error => panic!("error state"),
-        }
-    }
-
-    pub fn into_inner(self) -> PacketBodyReader<Box<dyn BufRead + 'a>> {
-        match self {
-            Self::Init { source, .. } => source.into_inner(),
-            Self::Body { source, .. } => source.into_inner(),
-            Self::Done { source, .. } => source.into_inner(),
-            Self::Error => panic!("error state"),
-        }
-    }
-
-    fn fill_inner(&mut self) -> io::Result<()> {
-        if matches!(self, Self::Done { .. }) {
-            return Ok(());
-        }
-
-        loop {
-            match std::mem::replace(self, Self::Error) {
-                Self::Init {
-                    mut hasher,
-                    mut source,
-                } => {
-                    debug!("SignatureOnePassReader init");
-                    let mut buffer = BytesMut::zeroed(1024);
-                    let read = fill_buffer(&mut source, &mut buffer, None)?;
-                    buffer.truncate(read);
-
-                    if read == 0 {
-                        return Err(io::Error::new(
-                            io::ErrorKind::UnexpectedEof,
-                            "missing signature",
-                        ));
-                    }
-
-                    // TODO: normalize line endings..
-                    hasher.update(&buffer[..read]);
-
-                    *self = Self::Body {
-                        source,
-                        hasher,
-                        buffer,
-                    };
-                }
-                Self::Body {
-                    mut hasher,
-                    mut source,
-                    mut buffer,
-                } => {
-                    debug!("SignatureOnePassReader body");
-
-                    if buffer.has_remaining() {
-                        *self = Self::Body {
-                            hasher,
-                            source,
-                            buffer,
-                        };
-                        return Ok(());
-                    }
-
-                    buffer.resize(1024, 0);
-                    let read = fill_buffer(&mut source, &mut buffer, None)?;
-                    buffer.truncate(read);
-
-                    // TODO: normalize line endings
-                    hasher.update(&buffer[..read]);
-
-                    if read == 0 {
-                        debug!("SignatureOnePassReader finish");
-
-                        let (reader, parts) = source.into_parts();
-
-                        // read the signature
-                        let mut packets = crate::packet::PacketParser::new(reader);
-                        let Some(packet) = packets.next() else {
-                            return Err(io::Error::new(
-                                io::ErrorKind::UnexpectedEof,
-                                "missing signature packet",
-                            ));
-                        };
-                        let packet = packet.map_err(|e| {
-                            io::Error::new(io::ErrorKind::InvalidData, e.to_string())
-                        })?;
-
-                        let Packet::Signature(signature) = packet else {
-                            return Err(io::Error::new(
-                                io::ErrorKind::UnexpectedEof,
-                                format!(
-                                    "missing signature packet, found {:?} instead",
-                                    packet.tag()
-                                ),
-                            ));
-                        };
-
-                        // calculate final hash
-                        let len =
-                            signature
-                                .config
-                                .hash_signature_data(&mut hasher)
-                                .map_err(|e| {
-                                    io::Error::new(io::ErrorKind::InvalidData, e.to_string())
-                                })?;
-                        hasher.update(&signature.config.trailer(len).map_err(|e| {
-                            io::Error::new(io::ErrorKind::InvalidData, e.to_string())
-                        })?);
-                        let hash = hasher.finalize();
-
-                        // reconstruct message source
-                        let reader = packets.into_inner();
-                        let source = parts.into_message(reader);
-
-                        *self = Self::Done {
-                            signature,
-                            hash,
-                            source: Box::new(source),
-                        };
-                    } else {
-                        *self = Self::Body {
-                            hasher,
-                            source,
-                            buffer,
-                        }
-                    }
-
-                    return Ok(());
-                }
-                Self::Done {
-                    hash,
-                    source,
-                    signature,
-                } => {
-                    *self = Self::Done {
-                        hash,
-                        source,
-                        signature,
-                    };
-                    return Ok(());
-                }
-                Self::Error => panic!("error state"),
-            }
-        }
-    }
-
-    pub fn is_done(&self) -> bool {
-        matches!(self, Self::Done { .. })
-    }
-
-    fn decompress(self) -> Result<Self> {
-        match self {
-            Self::Init { hasher, source } => {
-                let source = source.decompress()?;
-                Ok(Self::Init {
-                    hasher,
-                    source: Box::new(source),
-                })
-            }
-            _ => {
-                bail!("cannot decompress message that has already been read from");
-            }
-        }
-    }
-
-    fn decrypt(
-        self,
-        key_pws: &[Password],
-        keys: &[&SignedSecretKey],
-    ) -> Result<(Self, Vec<Fingerprint>)> {
-        match self {
-            Self::Init { hasher, source } => {
-                let (source, fps) = source.decrypt(key_pws, keys)?;
-                Ok((
-                    Self::Init {
-                        hasher,
-                        source: Box::new(source),
-                    },
-                    fps,
-                ))
-            }
-            _ => {
-                bail!("cannot decrypt message that has already been read from");
-            }
-        }
-    }
-
-    fn decrypt_with_password(self, msg_pw: &Password) -> Result<Self> {
-        match self {
-            Self::Init { hasher, source } => {
-                let source = source.decrypt_with_password(msg_pw)?;
-                Ok(Self::Init {
-                    hasher,
-                    source: Box::new(source),
-                })
-            }
-            _ => {
-                bail!("cannot decrypt message that has already been read from");
-            }
-        }
-    }
-}
-
-enum MessageParts {
+pub(crate) enum MessageParts {
     Literal {
         packet_header: PacketHeader,
         header: LiteralDataHeader,
@@ -529,7 +78,7 @@ enum MessageParts {
 }
 
 impl<'a> Message<'a> {
-    fn into_parts(self) -> (Box<dyn BufRead + 'a>, MessageParts) {
+    pub(crate) fn into_parts(self) -> (Box<dyn BufRead + 'a>, MessageParts) {
         match self {
             Message::Literal { reader } => {
                 assert!(reader.is_done());
@@ -620,7 +169,7 @@ impl<'a> Message<'a> {
 }
 
 impl MessageParts {
-    fn into_message<'a>(self, reader: Box<dyn BufRead + 'a>) -> Message<'a> {
+    pub(crate) fn into_message<'a>(self, reader: Box<dyn BufRead + 'a>) -> Message<'a> {
         match self {
             MessageParts::Literal {
                 packet_header,
@@ -660,45 +209,6 @@ impl MessageParts {
                 esk,
                 is_protected,
             } => todo!(),
-        }
-    }
-}
-
-impl BufRead for SignatureOnePassReader<'_> {
-    fn fill_buf(&mut self) -> io::Result<&[u8]> {
-        self.fill_inner()?;
-        match self {
-            Self::Init { .. } => panic!("invalid state"),
-            Self::Body { buffer, .. } => Ok(&buffer[..]),
-            Self::Done { .. } => Ok(&[][..]),
-            Self::Error => panic!("error state"),
-        }
-    }
-
-    fn consume(&mut self, amt: usize) {
-        match self {
-            Self::Init { .. } => panic!("invalid state"),
-            Self::Body { buffer, .. } => {
-                buffer.advance(amt);
-            }
-            Self::Done { .. } => {}
-            Self::Error => panic!("error state"),
-        }
-    }
-}
-
-impl Read for SignatureOnePassReader<'_> {
-    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        self.fill_inner()?;
-        match self {
-            Self::Init { .. } => panic!("invalid state"),
-            Self::Body { buffer, .. } => {
-                let to_write = buffer.remaining().min(buf.len());
-                buffer.copy_to_slice(&mut buf[..to_write]);
-                Ok(to_write)
-            }
-            Self::Done { .. } => Ok(0),
-            Self::Error => panic!("error state"),
         }
     }
 }
@@ -1295,7 +805,7 @@ impl<'a> Message<'a> {
         Ok(out)
     }
 
-    fn into_inner(self) -> PacketBodyReader<Box<dyn BufRead + 'a>> {
+    pub fn into_inner(self) -> PacketBodyReader<Box<dyn BufRead + 'a>> {
         match self {
             Self::Literal { reader } => reader.into_inner(),
             Self::Compressed { reader } => reader.into_inner(),
